backgrounds: four rooms an achievement pays out
The prize pool was eleven places and three ways to come by one — raise a
creature, clear an area, take a bracket. Every achievement ladder on the
device paid Bits and a cache and nothing you could look at.

BackgroundSource::Achieve is the fourth way, and it is the one that
scales: a row names an achievement id and Game::backgroundOwned asks the
bitset that is already on the save. Ownership stays derived, so this
costs the blob nothing and cannot disagree with itself. The announce is
cheaper than the other three — unlockAchievement's early return already
established the bit was clear, so there is no ask-before/ask-after to get
wrong.

Four places, each the ROOM its ladder is about:

  THE LINE        a kitchen pass, for knowing ten recipes. Stainless in
                  pure value, with the only tint in the scene at the
                  burners — a warm point of light in a grey room.
  THE CRT BENCH   the workshop the rig is built on, for a full rig.
                  Pegboard, a rank of tubes, two of them carrying a live
                  trace on frag-lo phosphor.
  GROUND STATION  a ridge of dishes for mapping the spectrum. The one
                  place in the set that is genuinely empty, which is what
                  a screen with a full banner on it needs.
  TRACE CITY      a board read as a skyline, for a hundred thousand
                  steps. Orthogonal routing rather than a receding grid,
                  because Grid Horizon is already that picture.

An Achieve row's picker line names the DEED and never the threshold —
that number lives on the achievement row and only there. The gate checks
the id resolves, so a renamed achievement is a build failure rather than
a background nobody can ever earn.

The picker's ownership mask goes to 32 bits. Fifteen rows would have sat
one under a uint16_t's ceiling, and the failure mode past it is rows
silently drawing LOCKED forever rather than a build that stops.

+++ src/core/ui/cfg_screen.cpp
@@ -243,14 +243,14 @@ void drawCfgDevice(Framebuffer& fb, int cursor, UiMode uiMode, int brightness,
     drawHintBand(fb, "A NEXT  B OPEN  C BACK");
 }
 
-void drawBackgrounds(Framebuffer& fb, int pick, uint16_t ownedMask, int equipped,
+void drawBackgrounds(Framebuffer& fb, int pick, uint32_t ownedMask, int equipped,
                      int beat) {
     drawHeaderBand(fb, "BACKGROUND");
 
     // One line of copy, and it belongs to whichever row is FOCUSED — which is what lets
-    // the list carry twelve rows without twelve hints in it, and what makes walking the
-    // locked ones worth doing. AUTO's line says what AUTO does; every other row's is
-    // its own `earnedBy`.
+    // the list carry a table's worth of rows without a hint per row, and what makes
+    // walking the locked ones worth doing. AUTO's line says what AUTO does; every other
+    // row's is its own `earnedBy`.
     const char* why = "MATCHES YOUR PET";
     if (pick > 0 && pick <= kBackgroundCount) why = kBackgrounds[pick - 1].earnedBy;
     drawText(fb, kMargin, 30, why, palColor(Pal::INK_DIM));
